Change how files from navbar list are handled
When the user clicked a filename from the navbar list, original behavior
was to erroneously go to the edit page, without any processing. This
would show the user a blank page.

Now, clicking a filename will first check if a folder path with
the same name as the file already exists. If it does, the site will
navigate to the edit page. If it doesn't, a message is printed to
console and the index page is refreshed.

This is only an issue while example files are stored on the server.
Other options are to remove the example files or only show directories
in the navbar list.

Ideally, a future iteration will allow processing of an example pdf by
simply clicking on it.